Add refresh masterData button in Result screen + move MasterData to App.js
