+++ App.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { LogBox, StyleSheet, useColorScheme } from "react-native";
 import { globalColors as GlobalColors } from "./shared/GlobalStyles";
 import { NavigationContainer } from "@react-navigation/native";
@@ -7,6 +7,10 @@ import { createMaterialTopTabNavigator } from "@react-navigation/material-top-ta
 import Dashboard from "./src/main/Dashboard";
 import Result from "./src/main/Result";
 import { SafeAreaProvider, SafeAreaView } from "react-native-safe-area-context";
+import { useToggle } from "./shared/hooks/useToggle";
+import AsyncStorage from "@react-native-async-storage/async-storage";
+import { AsyncStorageKeys } from "./src/storage/AsyncStorageKeys";
+import { DEBUG_MODE, USE_SAMPLE_DATA } from "./shared/GlobalConstants";
 
 const Tab = createMaterialTopTabNavigator();
 const App = () => {
@@ -22,6 +26,70 @@ const App = () => {
     backgroundColor: isDarkMode ? GlobalColors.grey : GlobalColors.white,
   };
 
+  // start configuring masterData
+  const [isLoading, toggleLoading] = useToggle(true);
+  const [masterData, setMasterData] = useState([]);
+
+  const loadData = async () => {
+    return JSON.parse(await AsyncStorage.getItem(AsyncStorageKeys.masterData));
+  };
+
+  const saveData = masterData => {
+    AsyncStorage.setItem(AsyncStorageKeys.masterData, JSON.stringify(masterData));
+  };
+
+  useEffect(() => {
+    // set timer loading UI
+    if (!DEBUG_MODE) {
+      setTimeout(() => {
+        toggleLoading(false);
+      }, 1000);
+    }
+
+    if (USE_SAMPLE_DATA) {
+      const setData = () => {
+        const sampleData = [
+          // { name: "hung vu", amount: 69000 },
+          // { name: "alice", amount: 25000 },
+          // { name: "xi peso", amount: 1200000 },
+          // { name: "johnny walker", amount: 1200000 },
+          // { name: "moon peso", amount: 1200000 },
+          // { name: "john doe", amount: 1200000 },
+          // { name: "vu hung", amount: 1200000 },
+          // { name: "george bush", amount: 1200000 },
+          // { name: "the queen of england", amount: 1200000 },
+        ];
+
+        saveData(sampleData);
+      };
+      setData();
+    }
+
+    // load masterData from local storage on startup
+    loadData().then(res => {
+      console.log("Startup::", res);
+      if (!!res) {
+        setMasterData(res);
+      }
+      toggleLoading(false)
+    });
+  }, []);
+
+  useEffect(() => {
+    if (!!masterData) {
+      console.log("masterData changes::", masterData);
+      setMasterData(masterData.sort((a, b) => b.amount - a.amount))
+      saveData(masterData);
+    }
+  }, [masterData]);
+
+  const dashboardProps = {
+    isLoading,
+    masterData,
+    setMasterData,
+    loadData,
+  };
+
   return (
     <SafeAreaView style={[backgroundStyle, styles.safeAreaViewContainer]}>
       <SafeAreaProvider>
@@ -33,13 +101,13 @@ const App = () => {
           >
             <Tab.Screen name={ROUTES.DASHBOARD_SCREEN}>
               {(props) =>
-                <Dashboard {...props} />
+                <Dashboard {...props} dashboardProps={dashboardProps}/>
               }
             </Tab.Screen>
 
             <Tab.Screen name={ROUTES.RESULT_SCREEN}>
               {(props) =>
-                <Result {...props} />
+                <Result {...props} dashboardProps={dashboardProps} />
               }
             </Tab.Screen>
           </Tab.Navigator>
