+++ src/main/Dashboard/index.js
@@ -3,12 +3,12 @@ import DashboardScreen from "./DashboardScreen";
 import React from "react";
 import { SafeAreaView } from "react-native";
 
-const Dashboard = ({ navigation }) => {
+const Dashboard = ({ navigation, dashboardProps }) => {
 
   return (
     <SafeAreaView style={{flex: 1}}>
       <DashboardHeader navigation={navigation} />
-      <DashboardScreen />
+      <DashboardScreen dashboardProps={dashboardProps}/>
     </SafeAreaView>
   );
 };
