+++ src/main/Result/index.js
@@ -1,11 +1,14 @@
 import React from "react";
 import ResultHeader from "./ResultHeader";
 
-const Result = ({navigation}) => {
+const Result = ({navigation, dashboardProps}) => {
 
   return (
     <>
-      <ResultHeader navigation={navigation}/>
+      <ResultHeader
+        navigation={navigation}
+        dashboardProps={dashboardProps}
+      />
     </>
   )
 }
